+++ DearPyGui/src/core/AppItems/values/mvValueRegistry.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "mvAppItem.h"
 #include "mvItemRegistry.h"
-#include "mvModule_Core.h"
+#include "mvModule_DearPyGui.h"
 #include "cpp.hint"
 
 namespace Marvel {