+++ Distribution/BuildPythonWheel.py
@@ -17,7 +17,7 @@ if not os.path.isdir(script_dir +  "/dearpygui/"):
 
 # copy add items to temporary location
 shutil.copy(location, script_dir +"/dearpygui")
-shutil.copy(script_dir + "/../DearPyGui/dearpygui/core.pyi", script_dir + "/dearpygui")
+shutil.copy(script_dir + "/../DearPyGui/dearpygui/_dearpygui.pyi", script_dir + "/dearpygui")
 shutil.copy(script_dir + "/../DearPyGui/dearpygui/dearpygui.py", script_dir + "/dearpygui")
 shutil.copy(script_dir + "/../DearPyGui/dearpygui/logger.py", script_dir + "/dearpygui")
 shutil.copy(script_dir + "/../DearPyGui/dearpygui/demo.py", script_dir + "/dearpygui")